multi_data_semantic_seg/src/requirements.txt
numpy>=1.20
Pillow>=8.0
scipy>=1.7
torch>=1.9
opencv-python-headless>=4.5
# mmsegmentation, mmengine, mmcv for inference (install separately if using mmseg models)
